--- crates/dbg-swc/src/bin/cmp-size.rs
@@ -1,5 +1,0 @@
-//! https://embed.cs.utah.edu/creduce/using/
-//!
-//! This binary is used to check if `input.js` is interesting,
-//! i.e. if terser or esbuild generates smaller output for input.js,
-//! it's interesting (in creduce terms) and this script returns 0
